# packages/engine/paima-rest/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"DryRunResponse": {
				"properties": {
					"valid": {
						"type": "boolean"
					}
				},
				"required": [
					"valid"
				],
				"type": "object"
			},
			"SuccessfulResult_DryRunResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/DryRunResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailedResult": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"errorMessage": {
						"type": "string"
					},
					"errorCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"errorMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_DryRunResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_DryRunResponse_"
					},
					{
						"$ref": "#/components/schemas/FailedResult"
					}
				]
			},
			"InternalServerErrorResult": {
				"$ref": "#/components/schemas/FailedResult"
			},
			"FieldErrors": {
				"description": "comes from the `tsoa` package, but we don't want it as a dependency just for this type",
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ValidateErrorResult": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation Failed"
						],
						"nullable": false
					},
					"details": {
						"$ref": "#/components/schemas/FieldErrors"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LatestProcessedBlockheightResponse": {
				"properties": {
					"block_height": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"block_height"
				],
				"type": "object"
			},
			"EmulatedBlockActiveResponse": {
				"properties": {
					"emulatedBlocksActive": {
						"type": "boolean"
					}
				},
				"required": [
					"emulatedBlocksActive"
				],
				"type": "object"
			},
			"SuccessfulResult_number_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_number_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_number_"
					},
					{
						"$ref": "#/components/schemas/FailedResult"
					}
				]
			},
			"DeploymentBlockheightToEmulatedResponse": {
				"$ref": "#/components/schemas/Result_number_"
			},
			"SuccessfulResult_boolean_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_boolean_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_boolean_"
					},
					{
						"$ref": "#/components/schemas/FailedResult"
					}
				]
			},
			"ConfirmInputAcceptanceResponse": {
				"$ref": "#/components/schemas/Result_boolean_"
			},
			"TransactionCountResponse": {
				"properties": {
					"gameInputs": {
						"type": "number",
						"format": "double"
					},
					"scheduledData": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"gameInputs",
					"scheduledData"
				],
				"type": "object"
			},
			"SuccessfulResult_TransactionCountResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/TransactionCountResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_TransactionCountResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_TransactionCountResponse_"
					},
					{
						"$ref": "#/components/schemas/FailedResult"
					}
				]
			},
			"TransactionContentResponse": {
				"properties": {
					"from": {
						"type": "string"
					},
					"inputData": {
						"type": "string"
					},
					"txIndex": {
						"type": "number",
						"format": "double"
					},
					"blockNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"from",
					"inputData",
					"txIndex",
					"blockNumber"
				],
				"type": "object"
			},
			"SuccessfulResult_TransactionContentResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/TransactionContentResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_TransactionContentResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_TransactionContentResponse_"
					},
					{
						"$ref": "#/components/schemas/FailedResult"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@paima/rest",
		"version": "1.0.0",
		"description": "Paima Engine REST, defining the REST endpoints for the Paima node",
		"contact": {
			"name": "Paima Studios"
		}
	},
	"paths": {
		"/dry_run": {
			"get": {
				"operationId": "DryRunGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_DryRunResponse_"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "gameInput",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/latest_processed_blockheight": {
			"get": {
				"operationId": "LatestProcessedBlockheightGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LatestProcessedBlockheightResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/emulated_blocks_active": {
			"get": {
				"operationId": "EmulatedBlockActiveGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmulatedBlockActiveResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/deployment_blockheight_to_emulated": {
			"get": {
				"operationId": "DeploymentBlockheightToEmulatedGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentBlockheightToEmulatedResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "deploymentBlockheight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/confirm_input_acceptance": {
			"get": {
				"operationId": "ConfirmInputAcceptanceGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConfirmInputAcceptanceResponse"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "gameInput",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "blockHeight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/transaction_count/blockHeight": {
			"get": {
				"operationId": "TransactionCountBlockHeight",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_TransactionCountResponse_"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "blockHeight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/transaction_count/address": {
			"get": {
				"operationId": "TransactionCountGet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_TransactionCountResponse_"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "blockHeight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/transaction_content/blockNumberAndIndex": {
			"get": {
				"operationId": "TransactionContentBlockHeight",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_TransactionContentResponse_"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerErrorResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailedResult"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "blockHeight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "txIndex",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}